+++ pages/admin/pendingProduct.php
@@ -1,17 +1,21 @@
 <?php
+	session_start();
+	if (isset($_SESSION['sessAdminId'])) {
+		$products = $product->find('status', 0);
 
-	$products = $product->find('status', 0);
+		if (isset($_GET['ppId'])) {
+			$products = $product->updateStatus('status', 'lotNumber', $_GET['ppId']);
+			header('Location:pendingProduct'); 
+		}
 
-	if (isset($_GET['ppId'])) {
-		$products = $product->updateStatus('status', 'lotNumber', $_GET['ppId']);
-		header('Location:pendingProduct'); 
+		if(isset($_GET['pId'])){
+			$products = $product->delete('lotNumber',$_GET['pId']);
+			header('Location:pendingProduct'); 
+		}
+		$title = "Product";
+
+		$content = loadTemplate('templates/adminPendingProduct.php',['products' => $products]);
 	}
-
-	if(isset($_GET['pId'])){
-		$products = $product->delete('lotNumber',$_GET['pId']);
-		header('Location:pendingProduct'); 
-	}
-	$title = "Product";
-
-	$content = loadTemplate('templates/adminPendingProduct.php',['products' => $products]);
+	else
+		include 'adminLogIn.php';
 ?>
