+++ pages/admin/product.php
@@ -1,12 +1,16 @@
 <?php
+	session_start();
+	if (isset($_SESSION['sessAdminId'])) {
+		$products = $product->find('status', 1);
 
-	$products = $product->find('status', 1);
+		if(isset($_GET['pId'])){
+			$products = $product->delete('lotNumber',$_GET['pId']);
+			header('Location:product'); 
+		}
+		$title = "Product";
 
-	if(isset($_GET['pId'])){
-		$products = $product->delete('lotNumber',$_GET['pId']);
-		header('Location:product'); 
+		$content = loadTemplate('templates/adminProduct.php',['products' => $products]);
 	}
-	$title = "Product";
-
-	$content = loadTemplate('templates/adminProduct.php',['products' => $products]);
+	else
+		include 'adminLogIn.php';
 ?>
